pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "A demo customer support agent built with FastAPI, SQLite, and React."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
]
llm = [
  "pydantic-ai-slim[openai]>=1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
